authenticated

+++ src/routers/plant.js
@@ -1,13 +1,18 @@
 const express = require('express')
 const Plants = require('../models/plant')
 const router = new express.Router
+const auth = require('../middleware/auth')
+
 
 router.get('/', (req,res)=>{
     res.send('hello')
 })
 
-router.post('/plant', async (req,res)=>{
-    const plant = new Plants(req.body)
+router.post('/plant',auth,  async (req,res)=>{
+    const plant = new Plants({
+        ...req.body,
+        owner: req.user._id
+    })
 
     try{
         await plant.save()
@@ -18,18 +23,22 @@ router.post('/plant', async (req,res)=>{
 
 })
 
-router.get('/plant', async (req,res)=>{
+router.get('/plant',auth, async (req,res)=>{
     try{
-        const plant = await Plants.find()
+        const plant = await Plants.find({owner: req.user._id})
         res.send(plant)
     }catch(e){
         res.status(400).send()
     }
 })
 
-router.get('/plant/:id', async (req,res)=>{
+router.get('/plant/:id',auth, async (req,res)=>{
+    const _id = req.params.id
+    
     try{
-        const plant = await Plants.findOne({_id: req.params.id})
+        //const plant = await Plants.findOne({_id: req.params.id})
+        const plant = await Plants.findOne({_id: _id, owner: req.user._id})
+
         if(!plant){
             res.status(404).send()
         }
@@ -39,9 +48,9 @@ router.get('/plant/:id', async (req,res)=>{
     }
 })
 
-router.delete('/plant/:id', async (req,res)=>{    
+router.delete('/plant/:id',auth, async (req,res)=>{    
     try{
-        const plant = await Plants.findOneAndDelete({_id: req.params.id})
+        const plant = await Plants.findOneAndDelete({_id: req.params.id, owner: req.user._id})
         if(!plant){
             return res.status(404).send()
         }
@@ -53,7 +62,7 @@ router.delete('/plant/:id', async (req,res)=>{
     }
 })
 
-router.patch('/plant/:id', async(req,res)=>{
+router.patch('/plant/:id',auth, async(req,res)=>{
   
     const updateArray = Object.keys(req.body)
     const updateTerms = ['name', 'water']
@@ -66,7 +75,7 @@ router.patch('/plant/:id', async(req,res)=>{
     }
     
     try{
-        const plant = await Plants.findOne({_id: req.params.id})
+        const plant = await Plants.findOne({_id: req.params.id, owner: req.user._id})
 
         if(!plant){
             res.status(404).send()
